% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{boot_biomass}
\alias{boot_biomass}
\title{Bootstrap the biomass estimates}
\usage{
boot_biomass(dat, reps = 1000)
}
\arguments{
\item{dat}{A data frame returned by \code{\link[gfdata:get_survey_sets]{gfdata::get_survey_sets()}}.}

\item{reps}{The number of bootstrap replicates.}
}
\description{
Bootstrap the biomass estimates
}
