% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{calc_bio}
\alias{calc_bio}
\title{Calculate a design-based biomass index}
\usage{
calc_bio(dat, i = seq_len(nrow(dat)))
}
\arguments{
\item{dat}{A data frame returned by \code{\link[gfdata:get_survey_sets]{gfdata::get_survey_sets()}}.}

\item{i}{A vector of indices to include. By default this includes all rows of
data. This is set up to work with bootstrapping via \code{\link[=boot_biomass]{boot_biomass()}}.}
}
\description{
Calculate a design-based biomass index
}
