% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_file.R
\name{clip_by_mpa}
\alias{clip_by_mpa}
\title{Clip survey set data by proposed MPA closed areas}
\usage{
clip_by_mpa(data = survey_sets, ssid)
}
\arguments{
\item{data}{Survey set data for a species from \code{gfdata::get_survey_sets()}}

\item{ssid}{Survey series id(s) to select closed areas}
}
\description{
Clip survey set data by proposed MPA closed areas
}
