% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_file.R
\name{closed_areas}
\alias{closed_areas}
\title{Create a spatial file of closed areas by fishery/survey type}
\usage{
closed_areas(
  mpa_network_shp = closed,
  ssid = NULL,
  fishery = NULL,
  level = c("X", "C"),
  out_crs = 3156
)
}
\arguments{
\item{mpa_network_shp}{MPA polygon shapefile (draft from gdb from Katie Gale Apr 2020); filter for desired restricted zones}

\item{ssid}{Survey series id to select restricted areas for}

\item{fishery}{Fishery type to select restricted areas for (eg. "trawl", "ll" (longline), "trap" (not yet implemented))}

\item{level}{Level of restriction. "X" = closed to all activity of associated fishery/survey type, "C" = closed on certain conditions}

\item{out_crs}{Coordinate reference system for the output polygon}
}
\description{
Create a spatial file of closed areas by fishery/survey type
}
