% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_file.R
\name{design_biomass}
\alias{design_biomass}
\title{Survey design-based biomass index}
\usage{
design_biomass(dat)
}
\arguments{
\item{dat}{List containing survey set data by species (alread filtered for desired survey series id(s))}
}
\description{
Survey design-based biomass index
}
