% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusion.R
\name{exclude_areas}
\alias{exclude_areas}
\title{Exclude areas from the survey domain}
\usage{
exclude_areas(data, exclude_poly, X, Y)
}
\arguments{
\item{data}{A data frame from \code{\link[gfdata:get_survey_sets]{gfdata::get_survey_sets()}}.}

\item{exclude_poly}{FIXME:}

\item{X}{FIXME:}

\item{Y}{FIXME:}
}
\description{
Filter output of \code{\link[gfdata:get_survey_sets]{gfdata::get_survey_sets()}} to exclude sets within a given set of polygons.
}
