% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_file.R
\name{import_survey_sets}
\alias{import_survey_sets}
\title{Load survey set data}
\usage{
import_survey_sets(spp, ssid, dir, min_year = 1950)
}
\arguments{
\item{spp}{Species name(s) to get survey data for (in lower case; e.g., spp = "yelloweye rockfish" or spp = c("yelloweye rockfish", "pacific cod"))}

\item{ssid}{Survey series id(s) to get survey data for}

\item{dir}{Directory where species data from gfdata::get_survey_sets() exists (.rds file(s))}

\item{min_year}{Earliest year to retrieve survey data for}
}
\description{
Load survey set data
}
