% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_file.R
\name{plot_survey_pts}
\alias{plot_survey_pts}
\title{Plot survey point data}
\usage{
plot_survey_pts(data = survey_sets, ssid = NULL)
}
\arguments{
\item{data}{Data for one species from \code{gfdata::get_survey_sets()} (or \code{import_survey_sets()[spp]})}

\item{ssid}{Survey series id(s) to plot data for}
}
\description{
Plot survey point data
}
