% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_file.R
\name{sdmTMB_biomass}
\alias{sdmTMB_biomass}
\title{Survey st model-based biomass index}
\usage{
sdmTMB_biomass(dat)
}
\arguments{
\item{dat}{List containing survey set data by species (alread filtered for desired survey series id(s))}
}
\description{
Survey st model-based biomass index
}
